v1.21 * Fixed some command messages * Added more `delay:` options to `/do` * Added `/rtp` *- randomly teleport a player (staff use only)* * Added `/seenlast` oshowing when someone was last promoted * Added `/setgroup` command * Added macros and variables! *More info on this soon...* * Fixed using quotes and apostraphes in command that use quotes such as `/raw` * Changed `/promote`'s color to #79ea75

+++ src/main/java/com/akoot/plugins/ultravanilla/commands/TempBanCommand.java
@@ -26,7 +26,7 @@ public class TempBanCommand extends AdminCommand implements CommandExecutor {
             OfflinePlayer target = plugin.getServer().getOfflinePlayer(args[0]);
             if (UltraVanilla.getConfig(target.getUniqueId()) != null) {
                 if (!target.isBanned()) {
-                    long time = StringUtil.getSeconds(args[1]) * 1000L;
+                    long time = (long) (StringUtil.getSeconds(args[1]) * 1000.0);
                     long now = System.currentTimeMillis();
                     long expires = now + time;
                     String reason = getArg(args, 3);
